requirements.txt
flask
python-dotenv
line-bot-sdk==3.21.0
openai